requirements.txt
pandas==2.2.3
python-dateutil==2.9.0.post0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
fastapi==0.115.12
uvicorn==0.34.0
supabase==2.15.3
